refactor(simulation): split create web mapper
